+++ discovery-provider/plugins/pedalboard/apps/relay/src/web3.ts
@@ -1,6 +1,9 @@
 import { ethers } from "ethers";
 import { Config } from "./config/config";
 import { logger } from "./logger";
+import Web3 from "web3";
+
+export let web3js: Web3
 
 export type Web3Result = {
   web3: ethers.providers.JsonRpcProvider;
@@ -14,6 +17,7 @@ export const connectWeb3 = async (config: Config): Promise<Web3Result> => {
     // TODO: check nethermind specific health here
     try {
       const web3 = new ethers.providers.JsonRpcProvider(rpcEndpoint);
+      web3js = new Web3(new Web3.providers.HttpProvider(rpcEndpoint))
       const { chainId } = await web3.getNetwork();
       return { web3, chainId: chainId.toString() };
     } catch (e) {
@@ -27,5 +31,24 @@ export const connectWeb3 = async (config: Config): Promise<Web3Result> => {
   console.warn(`falling back to ${rpcEndpointFallback}`);
   const web3 = new ethers.providers.JsonRpcProvider(rpcEndpointFallback);
   const { chainId } = await web3.getNetwork();
+  web3js = new Web3(new Web3.providers.HttpProvider(rpcEndpointFallback))
   return { web3, chainId: chainId.toString() };
+};
+
+export const confirm = async (
+  txHash: string,
+  retries = 24
+): Promise<any> => {
+  let tries = 0;
+  while (tries !== retries) {
+    const receipt = await web3js.eth.getTransactionReceipt(txHash);
+    if (receipt !== null && receipt.status) return receipt;
+    await delay(250);
+    tries += 1;
+  }
+  throw new Error(`transaction ${txHash} could not be confirmed`);
+};
+
+const delay = (ms: number) => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
 };
